refactor: 회원가입 types 수정정
