+++ api/signup.ts
@@ -1,6 +1,6 @@
 import apiClient from './apiClient';
-import { SignUpRequest } from '@/types/SignUpRequest';
-import { SignUpResponse } from '@/types/SignUpResponse';
+import { SignUpRequest } from '@/types/SignUp';
+import { SignUpResponse } from '@/types/SignUp';
 
 export const signUp = async (data: SignUpRequest): Promise<SignUpResponse> => {
     const response = await apiClient.post<SignUpResponse>('/users/signup', data);
