--- types/SignUpResponse.ts
@@ -1,2 +1,0 @@
-export interface SignUpResponse {
-}
